ai_safety_service/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
httpx
pytest
pytest-asyncio
python-multipart
redis
celery
transformers
torch
numpy
pandas
scikit-learn
pillow
python-jose[cryptography]
passlib[bcrypt]
python-magic
aiofiles
tenacity
nltk
textblob
wordnet

# Additional dependencies for the full AI Safety Service
openai
tiktoken
sentence-transformers
langchain
langchain-community
langchain-core
evaluate
adversarial-robustness-toolbox
fairlearn
aiohttp
asyncio
backoff
bert-score
rouge
bleach
matplotlib
seaborn
huggingface-hub
datasets
torchvision
torchaudio
dacite
jsonschema
attrs
regex 